Fix the bogus trace_break_arg constructed at trace error breakpoints
When the errorcatch trace option was set, a bogus trace_break_arg was
being constructed when choosing to retry with new arguments.  Here
trace_break_arg is supposed to be bound to a (Maxima) list of the
previous arguments, but instead it was bound to a singleton (Maxima)
list with a lisp list of previous arguments as its element.

This easily led to lisp errors when attempting to use it.

This bug has been around since at least as far back as Maxima v5.0.