Add support for multiple return values to the ERRSET macro
The previous definition of the ERRSET macro (which is like the ERRSET
special form in Maclisp) was only returning the primary return value
of the evaluated form in the non-error case.

Now a list of all return values is returned in the non-error case (but
there is one ambiguous case which I describe in a new comment).

The test suite runs fine.