Fix bug #2803: trace/timer do not work for functions with quoted arguments
Functions with one or more quoted formal arguments (mfexprs) were
not behaving correctly when traced or timed.  The problem was that
usual function argument evaluation rules were being used instead
of the special evaluation rules required for these functions.

Now instead of putting the hook function in the symbol's function
cell like for usual mexprs, we put a (Maxima) lambda expression on
the symbol's plist using the key MFEXPR.  This lambda expression is
constructed so that the appropriate evaluation rules are used for
the function's arguments before calling the hook.

mfexpr*s also use the symbol's plist, but the hook goes directly
on the plist.  That works because the evaluator normally looks
there for mfexpr*s anyway, and the evaluation rule is simple.
I'm wrapping the hook for mfexprs in a lambda expression so we
can just call MLAMBDA in the evaluator instead of doing the
special evaluation of the arguments and calling the hook
separately ourselves.

The trace-type for these functions is still MEXPR because the
rest of the trace machinery works fine without introducing a
new trace-type.

The test suite runs fine, with new tests.