Fix bug #3439: Simplode output on empty list
The problem was that calling simplode on an empty list and with
a nonempty delimiter would return the string "false" instead of
the empty string "".

Now just punt to sconcat by calling it with no arguments in the
case of an empty list.  This gives the expected empty string return
value and it is consistent with the code comment that simplode is
"sconcat for lists".

The test suite and share test suite run fine, with new tests.