An idea on how we can test draw automatically.
We cannot automatically check if the result gnuplot gives us
makes any sense. Well, Google's KI might, perhaps. And the tests/
folder contains a few files that generate plots one can check
manually. But we can test in an automated way if gnuplot's input
contains the lines we believe it to contain.

This test suite additionally checks if draw can be loaded at all.
There have been problems with this in the past...