The daily draw test enhancements and a small bugfix for draw
This commit adds tests for draw options beginning with the
letter "g". And the fix for a small bug that was found during
testing:

It makes sense to allow the user to specify the grid frequency
as a floating-point number: 1.0 or 2.0 obviously are valid
frequencies - but floats.
But it does not make too much sense to actually request things
like 5.2 lines between two major axis ticks from gnuplot.
My personal version of gnuplot seems to round the number down
to 5 butas there are many gnuplot versions around one can
expect one to fail to do that

 => Now maxima rounds the grid parameters before sending them
to gnuplot.