On sbcl run_testsuite(share_tests=true) did sometimes OOM
Let's try to prevent the out-of-memory by manually triggering a
garbage collection on loading a package.

The speed penalty of this garbage collection is low as it appears
only seldomly. Let's see, though, if the build server tells us
that it actually helps.

If it doesn't help I will revert this commit.
My guess is that tweaking sb-ext:bytes-consed-between-gcs might
reduce the memory fragmentation by sbcl that is the main reason
for running out of memory. But that would be a bigger change so
I'd like to try it last.