Bug fix: MEVALP_TR error messages always said the predicate was false
If MEVALP_TR could not evaluate the predicate and it was asked to
signal an error, the error message was always

  Unable to evaluate predicate false

It looks like this has been broken since commit 3fb19813 (11 years
ago) when PATEVALLED was no longer declared special.  Prior to that
commit, MEVALP_TR used the value of PATEVALLED as set in MEVALP1_TR
for error messages.  A few weeks after that commit it was noticed
that PATEVALLED was not bound in MEVALP1_TR, so a lexical binding
was added there in commit 1eec31de.

This commit reverts 1eec31de and declares the binding and references
to PATEVALLED to be dynamic in MEVALP_TR and MEVALP1_TR.

The test suite and share test suite both run fine.