Fix bug #3410: extremal_subset gives bogus result with undefined predicate
This applies the fix from Stavros Macrakis in the bug report.
Now MEVALP_TR is used directly instead of IS-BOOLE-CHECK in
extremal_subset and we explicitly have MEVALP_TR signal an
error when it can't evaluate the predicate.

The test suite and share test suite both run fine, with a new test.