Provide the system with a way to tell maxima to start in a certain directory.
On MS Windows maxima needs to be started in the directory it is installed
in as it needs the current working directory for determining where to
find gnuplot and where to find all the files belonging to maxima.

If maxima needs to access files from the directory a file
(for example a wxMaxima worksheet) was opened in some lisps provide a
way to change the current working directory. But as common lisp doesn't
have a concept of a parent working directory these ways are neither
standard nor not riddled with bugs:
clisp doesn't like short path names, sbcl errors out on names that
contain non-UTF8 non-ASCII characters...

=> Added an environment variable named MAXIMA_PWD that tells the maxima
batch file/shell script which dir to change into before maxima is
started.

For the wxMaxima commits related to this feature see:
 - https://github.com/andrejv/wxmaxima/commit/acf527a8ae8eae1f7ded03012756deb007b0c855
 - https://github.com/andrejv/wxmaxima/commit/0f58038222af5cfb1e47d8738cc5f02253f765fe