Tweak the memory management for GCL towards big maxima projects
GCL before starting the garbage collector currently by default tries
to allocate too much of the system's physical memory to let a
firefox live in it and sometimes runs out of memory on easy tasks
if only maxima and a graphical user environment is running =>

I've told it to be less greedy by default. On my computer it still
was able to use all of the system's RAM if it wanted to.

Now an user has reported that this isn't the case for him => let's
try if tweeking the allocator a little bit more is more of a good
idea than not tweeking it further.