Build system: Un-weirdify the "make check" target
"make check" made use of sed and a .in.in file that
really was a template to a template.

After suggestions from Viktor Toth and Leo Butler we
now have a template for a shell test script that supports
one command-line parameter (the name of the lisp to test)
and one 2-line script per lisp that is auto-generated
but is simple enough for not to require sed and a template.