Windows installer: improve the workaround for lapack & sbcl
Instead of patching maxima.bat.in during the build of the 64 bit
installer (where the patch must be adjusted from time to time, that
it applies cleanly), check if this is a 64 bit build by checking
the name of the installed Mingw libgcc (it is different for 32/64 bit).

Checking the processor architecture would not be enough, one could
install the 32 bit installer in a 64 bit operating system.