Work around ARRAY issue for ecl
Stick in (in-package :common-lisp") before running mk:oos.  This is
the same work-around used for lapack.  This gets rid of the compiler
error where ecl thinks a declaration is (maxima::array ...) instead of
(cl::array ...).  Don't know why this is, but I suspect this is a bug
in ecl's handling of packages.  No other Lisp compiler has this
problem.

Tested this with ecl 16.1.3 on OSX 10.13.  load("demo-dlsode") runs
and produces the expected output.

Also indented a line neatly and used maxima-load-pathname-directory
instead of using *load-pathname* or sys:*load-pathname*.