Replace defmfun with defmfun-checked
Comment out the original defmfun and rename defmun-checked to defmfun.

Needed to modify defmfun not to check for a leading $ for the function
name because maxima currently uses defmfun for many functions not
starting with $.  Just print a warning instead.  (Maybe remove that
eventually?)

Had to change maxima-display to be a defun instead of a defmfun
because it uses a keyword parameter which the new defmfun doesn't
support.  maxima-display isn't a exposed maxima function, so I think
this is harmless.