Replace defun with defmfun for user functions
This was basically done with the script

for f in `find src -name '*.lisp'`
do
  sed -i .bak 's;^(defun \$;(defmfun $;' $f
done

Bt for this to work, we need to load defmfun-check.lisp before
commac.lisp (that contained the old defmfun macro).

Also took the opportunity to move $-read-aux before x$-cl-macro-read
in commac.lisp so that $-read-aux is defined before being used in
x$-cl-macro-read.