Handle maclisp narg syntax even for user functions
share/tensor/itensor.lisp uses the syntax

(defmfun $defcon n ...)

This means we want to use the old maclisp narg syntax, and the defmfun
macro was expecting functions that begin with $ weren't use that.
Adjust the macro so that this is supported again.

Ran the entire test suite (including share tests) and all tests
passed. (Only tested with cmucl since the tests take a long time.)