Make the lists of testsuite files available directly on start of maxima
testsuite_files and share_testsuite_files were in the list of builtin
commands for rmaxima and also part of the documentation. But there was
no hint on both variables not being maxima lists and not being
available on startup but only after the testsuite was used or loaded
manually.

Now that both files actually are maxima lists the most natural thing
to do now felt to make the two variables defined on start-up of
maxima =>

 - the contents of the two files that defined the two variables has
   now been moved to src/testsuite.lisp in order to make it part
   of the source that is compiled.
 - maxima.system now compiles the new file into maxima's I/O module
 - the commands in mload.lisp that autoloaded the testsuite files
   were removed.
 - and a new README was added to the tests/ subdirectory informing
   about the new organization of the test suite.