Revert ""make check": Tried to make more lisp errors produce useful output"
ECL seems not to enter the debugger but on some machines on launchpad
the test bench seems to actually run for > 6 hours without proucing
any output.

My guess is that these machines are very low in RAM and therefore
start swaping at some point which can be a lengthy process. But
it's only a guess and might not be good as on my private machine
ECL only needs a handful of Megabytes for doing all tests...

...if anyone had a better idea I would be grateful.

This reverts commit 7af4b7badffb6d8357729e3368c6aab958a959f6.