Cleanup an example
One of the examples had the output from each individual command in a
separate @example/@end example section. This fix changes this to make
them all part of a single @example and group the individual commands
using @group/@end group. This is in line with all other examples in
the other files.