Let first arg of create_index be the file with the ToC
Using maxima_singlepage.html causes weird output to contents.hhc.  So
let's create contents.hhc from the file that contains the table of
contents.  This means create_index takes a single arg for that file.
This file is determined by running grep to find the one file that has
the table of contents.

Update Makefile to find the file and give it to create_index.

Also, with the change to use makeinf/texi2any, maxima.html isn't
created anymore, but maxima_0.html is.  So update the Makefile to use
maxima_0.html instead oof maxima.html.