share testsuite: Test for a mnewton failure that was resolved in 2010
On Jan 09, 2019 on he maxima-discuss list a discussion started
why genmatrix() is slow on sbcl. The reason was a
coerce_float_fun causing the creation of one function per
matrix element that was compiled my maxima.

the coerce_float_fun tried to resolve a problem with
mnewton that should be the one described in
https://def.fe.up.pt/pipermail/maxima-discuss/2010/033075.html

=> Added a test for this problem so we know if it recurrs whilst
resolving the performance issue.