Repaired the possibility to disable GCL's memory pool.
GCL by default tries to save CPU time by postponing garbage collects
until a selectable fraction of the system memory is used by the
current process - which means that a try to run two concurrent
maxima processes quickly ends in an out-of-memory.

=> maxima by default tells concurrent gcl processes to share
the information which percentage of the system memory gcl uses.

Unfortunately for the said option not only "Y" and "1" enable the
feature but also a "0" or a "N" => Martin Saturka proposed to
introduce an separate environment variable that disables the
shared memory pool in order to avoid any bug it contains.

Resolves #3529