"make clean" now cleans builds for currently-disabled lisps
If a lisp currently wasn't enabled when calling "./configure"
"make clean" didn't remove the remains from an eventual maxima
build for this lisp. Which means when the lisp was re-enabled
after upgrading the lisp or any dependency of it the next
build was prone to fail even if a "make clean" had been
issued inbetween.

Resolves the issue from the 20190308 maxima-discuss thread
"GCL build broken".