Bug fix: SIMPEXPT was marking unsimplified expressions with SIMP
Barton Willis reported some unsimplified expressions on the mailing
list.  A simpler case was found by Stavros Macrakis, who also found
the problem and provided the original fix.  RJF suggested a small
change to Stavros' fix.

The issue is that SIMPEXPT was assuming that certain powers it
constructed were simplified when that need not be the case.

The test suite and share test suite both run fine.