Fix an internal error in MSIZE-ARRAY (used in grind / 1D printing)
MSIZE-ARRAY would barf if given an MQAPPLY expression with an
operator that was not a cons.

One easy way to trigger this was with arraymake(1,[2]) with display2d
set to false.  This would cause SIMPLIFYA to try to print an error
message, which would cause the error in MSIZE-ARRAY, which would in
turn cause errormsg to complain:

(%i1) display2d : false$
(%i2) arraymake (1, [2]);
** error while printing error message **
simplifya: I don't know how to simplify this operator: ~M

This causes no problems in the test suite or share test suite.