Fix non-scalar tests in LENIENT-EXTENDED-REALP and LENIENT-COMPLEX-P
LENIENT-EXTENDED-REALP in src and LENIENT-COMPLEX-P in the
hypergeometric share package both had a featurep test to check
for the nonscalarp property instead of the intended nonscalar
property.

If a user used declare(nonscalarp,feature) to declare a new
feature with this name, then this would alter the behavior of
these functions in unanticipated ways.

No problems with the test suite or share test suite.