Mark rtest6/step 43 as "known bad" for ECL.
Don't know what has triggered this. But on the mailing list
nobody was able to find out what we have changed and it is
ECL-only => marking this step as "known bad" for the moment.

Running tests in rtest6:
********************** Problem 43 (line 166) ***************
Input:
(string(19.999999999999999e-8), (%% = 2.0e-7) or (%% = 2.0E-7) or %%)

Result:
19.999999999999999e-8

This differed from the expected result:
true