Remove unused ERRSET macro definition (depended on ECLDEBUG feature)
There have been two definitions of the ERRSET macro: one was used if
the ECLDEBUG feature was present and the other if not.  This was
introduced in 2008 in commit 7170f934 when attempting to get Maxima
with work with ECL (that's what I get from the commit message anyway).

The ECLDEBUG feature is not used or mentioned anywhere else.  Let's
just remove this duplication since this special case is unnecessary.
At this point, if we really want a separate ERRSET definition for
extra debugging then it shouldn't be specific to ECL.