Remove superfluous arguments passed to ERRSET
The ERRSET special form in Maclisp took an optional second argument
that specified whether or not to print the error message if an error
occurred inside the ERRSET.  Maxima's ERRSET macro has accepted a
second argument that was originally used for the same purpose (well,
sometimes: it depended on the underlying lisp).

The second argument to ERRSET has always been ignored since the port
to CL (and even before that on some pre-CL lisps).  ERRSET code for
pre-CL lisps was removed in 2004 in commit ad12e542.

I'm removing these extra arguments to ERRSET since it's misleading
and unnecessary to pass them when they're completely unused.

The ERRSET macro itself has not been changed, and it still accepts
multiple arguments for backward-compatibility with any code that
passes a second argument.

No functional changes.