Remove the unused special variable LISPERRPRINT
This special variable was sometimes passed as the second argument
to ERRSET with the intention of specifying whether or not to print
the error message if an error occurred within the ERRSET.  See the
commit message for commit 3ac79384 for more on this.

Since commit 3ac79384 this variable is unused in src and share
(strictly speaking, this variable was unused long before that; now
the symbol just does not appear anywhere in src or share).

If there is user code that uses (ERRSET FORM LISPERRPRINT) then it
will still work without change since ERRSET accepts more than one
argument and only uses FORM.