Now only bind ERRSET to T or NIL (do not reference unused handlers)
In Maclisp, if the value of ERRSET was not NIL then it specified a
handler for handling the condition caught by the ERRSET special form.

The ERRSET macro in Maxima for modern CL implementations has never
supported this.  Looking back in Maxima 5.0, it appears that the
ERRSET macros for Franz and Lucid did not support these handlers
either (they appear to only test whether or not value of ERRSET is
NIL).

It seems misleading and unclear to bind ERRSET to a symbol naming a
handler when the handler is never used.  Now instead of binding
ERRSET to a symbol naming an unused handler, ERRSET is simply bound
to T instead.  This does not affect the behavior, but the meaning of
the code is clearer.  Binding ERRSET to only T or NIL also matches
the comments in generr.lisp which describe the desired behavior of
ERRSET.

See also commit 3ac79384 for more on the differences between the
original Maclisp ERRSET special form and Maxima's ERRSET macro.

No functional changes.