Remove two unused functions and an unused special variable
The functions ERRBREAK1 and ERRBREAK2 were originally used as handlers
to be called when ERRSET caught an error.  These have not been used as
handlers since Maxima was ported to CL.  (See the commit message for
commit 0266e6e7 for more on this).  ERRBREAK1 would simply return NIL
and ERRBREAK2 would conditionally enter a break loop.

It seems that ERRBREAK1 was called directly in MERRSET when run under
the old pre-CL lisp NIL, but even that code (which was already unused
at the time) was removed in 2007 in commit 523080cf.

The special variable ERRBRKSW was originally used in ERRBREAK2 as a
flag to determine whether or not go into a break loop.

ERRBREAK1 and ERRBREAK2 are not called in src or share, and ERRBRKSW
is not used in src or share.  I'm removing these.