Remove some superfluous text from some break messages
tl;dr this removes an extra " T" from the end of two break messages;
they were added to the messages in 2007 for no apparent reason

In Maclisp there was a *BREAK function which took two arguments: a
flag to determine whether or not to go into a break loop, and then
a message to display.  There was also a BREAK special form which
had a different argument convention.

In Maxima 5.0 there was a *BREAK macro which was available if the
Multics feature was present.  It appears to be a wrapper around the
Maclisp (or similar) BREAK.

In 2007 in commit ecc752d4 the uses of *BREAK in Maxima were
rewritten to use the BREAK from CL, but the intended meaning of the
forms changed: the flag which simply specified whether or not to go
into a break loop became a format argument for the message.  So when
(*BREAK T ...) was originally used to unconditionally go into a break
loop, after the rewrite an extra "T" (actually, " T") was added to
the end of the break messages for no reason.

The break messages themselves could be improved, but there are
already other messages that get displayed before the break messages.
I'm removing the extra " T" from the end of the break messages since
they serve no purpose and were not intended to begin with anyway.