Maxima-clisp now should work again with front-ends
Maxima connects to its front-ends using a network socket.
And clisp seems to be somehow intelligent in trying to reduce the
network overhead by determining if it is better to merge several
strings maxima outputs to one network packet or if it is better
to split strings into several network packages in order to reduce
the latency.

With the newest version of wxWidgets and perl Xmaxima and
wxMaxima have begun to read the data from maxima faster than
maxima can produce data. Somehow this together with clisp's
intelligent output buffer triggers a bug that makes clisp
send the network packets in the wrong order. Which might be
that a long string maxima sends starts with the right 128
characters, then skips 128 characters and then finishes with
the characters that were skipped in the middle of a string.

Replacing all force-output commands by a finish-output seem to
somehow ship around this clisp bug. Hope none of the lisp wizards
finds out this is being a show stopper in a different place.