Remove some unused and undocumented translator variables
These variables are undocumented and unused in src and share:

  tr_gen_tags
  tr_output_file_default
  tr_predicate_brain_damage
  tr_version

The documentation for these was removed in 2005 in commit 5e5ba872.

tr_gen_tags was only used in pre-CL lisps.  The old pre-CL code that
used this variable was removed in 2007 in commit fe16647d, and other
related code was removed in 2011 in commit 0a370af9.

Some old code that had originally used tr_output_file_default was
removed in 2011 in commit 0a370af9.

tr_predicate_brain_damage was documented and defined but unused in
Maxima 5.0.

tr_version is always false by default.  Its value depended on the
VERSION property of the TRANSL-AUTOLOAD symbol.  In Maxima 5.0 it
appears that the VERSION property was only set while on a PDP-10.
There was also some commented-out code in Maxima 5.0 that set the
VERSION property.  The last of the old code that used TRANSL-AUTOLOAD
stuff was removed in 2011 in commit 11be1c1f.

This commit removes the definitions of these variables, as well as
references to them in emaxima (used for style and syntax highlighting
purposes).