Remove an incorrect comment
The comment states that the only *ARRAY calls remaining have a
second argument of NIL, but that's not true.  In CPBQ1 there is a
call with a second argument of T.

This comment was present in Maxima 5.0 and it was incorrect then
as well.  In fact, there were many more *ARRAY calls with a non-NIL
second argument then than there are now.

No functional changes.