Emaxima now correctly expands tabs in the comint buffer of emacs.
SF Patch #88 from the SF user Fede.
The message the patch came with is:

I installed maxima 5.42.2 from
https://sourceforge.net/projects/maxima/files/Maxima-Windows/5.42.2-Windows/maxima-clisp-sbcl-5.42.2-win64.exe/download
on Windows 10 with TeXLive 2018.

When I create a literate latex document with emaxima, if a referred cell is several lines long, the maxima cell output includes spurious characters. Please see the attached emaxima-lit.tex / .dvi files.

The problem is related to tabs. Referring to the attached patch. When emaxima inserts a command string in the maxima buffer it does include tabs. The region of the buffer is then processed by comint and as a side effect it replaces tabs with spaces. This results in that a previously registered position becomes incorrect.

Tha attached patch solves this problem.
Regards,
Fede