Remove the unused UUOLINKS property from plists
In Maclisp, so-called UUO links could be created to increase the
efficiency of function lookups in compiled functions.  The UUOLINKS
sstatus option could be used to "unsnap" all of the UUO links in the
environment.  This was used when a compiled function was redefined
or traced.

Franz lisp had a similar concept of transfer tables and a similar
option called TRANSLINK.

In Macsyma, the UUOLINKS property was used to unsnap the UUO links
or clear the transfer tables for subrs, lsubrs and fsubrs when they
were traced so that the new (traced) definitions were used instead
of the original (untraced) definitions.

The old Maclisp and Franz lisp code that used the UUOLINKS property
was removed in 2007 in commit abd4c9d8.